test(control-flow): isolate ON ERROR dispatch fixtures
