+++ CHANGELOG.md
@@ -3,6 +3,11 @@
   The existing executable, invocation order, assertions, and runtime behavior
   remain unchanged.
 
+- 2026-08-21: Continued #2564 test-module refactoring by moving the two
+  `ON ERROR DO` dispatch and metadata regressions into their own ownership
+  module. The aggregate `test_prg_engine_control_flow` target, invocation
+  order, assertions, and runtime behavior are unchanged.
+
 - 2026-08-21: Began the #2564 test-module refactor by moving the owned
   caught-runtime-fault/RETRY fixture into
   `test_prg_engine_control_flow_fault_recovery.cpp` while preserving its
