+++ tests/test_prg_engine_control_flow_on_error_dispatch.cpp
@@ -1,0 +1,129 @@
+// Copyright © 2026 Richard M. Hamilton.
+// SPDX-License-Identifier: GPL-3.0-only
+// Additional permission: Copperfin Application, Runtime, and Toolchain Exception 1.0; see LICENSE.
+
+#include "test_prg_engine_control_flow_support.h"
+
+namespace cf_test_prg_engine_control_flow {
+
+void test_on_error_do_handler_dispatches_routine() {
+    namespace fs = std::filesystem;
+    const fs::path temp_root = fs::temp_directory_path() / "copperfin_prg_engine_on_error_do";
+    std::error_code ignored;
+    fs::remove_all(temp_root, ignored);
+    fs::create_directories(temp_root);
+
+    const fs::path main_path = temp_root / "on_error_do.prg";
+    write_text(
+        main_path,
+        "ON ERROR DO handleerr\n"
+        "DO missing_routine\n"
+        "after_error = 1\n"
+        "RETURN\n"
+        "PROCEDURE handleerr\n"
+        "PUBLIC handled\n"
+        "handled = 1\n"
+        "RETURN\n");
+
+    copperfin::runtime::PrgRuntimeSession session =
+        copperfin::runtime::PrgRuntimeSession::create(make_runtime_session_options(main_path.string(), temp_root.string(), false));
+
+    const auto state = session.run(copperfin::runtime::DebugResumeAction::continue_run);
+    expect(state.completed, "ON ERROR DO script should complete after handler dispatch");
+
+    const auto handled = state.globals.find("handled");
+    const auto after_error = state.globals.find("after_error");
+    expect(handled != state.globals.end(), "ON ERROR handler should set handled flag");
+    expect(after_error != state.globals.end(), "execution should continue after ON ERROR handler returns");
+    if (handled != state.globals.end()) {
+        expect(copperfin::runtime::format_value(handled->second) == "1", "handled flag should be 1");
+    }
+    if (after_error != state.globals.end()) {
+        expect(copperfin::runtime::format_value(after_error->second) == "1", "post-error statement should run");
+    }
+
+    const bool has_handler_event = std::any_of(
+        state.events.begin(),
+        state.events.end(),
+        [](const copperfin::runtime::RuntimeEvent& event) {
+            return event.category == "runtime.error_handler";
+        });
+    expect(has_handler_event, "runtime should emit runtime.error_handler event when ON ERROR handler is dispatched");
+
+    fs::remove_all(temp_root, ignored);
+}
+
+void test_on_error_do_with_handler_receives_error_metadata() {
+    namespace fs = std::filesystem;
+    const fs::path temp_root = fs::temp_directory_path() / "copperfin_prg_engine_on_error_do_with";
+    std::error_code ignored;
+    fs::remove_all(temp_root, ignored);
+    fs::create_directories(temp_root);
+
+    const fs::path main_path = temp_root / "on_error_do_with.prg";
+    write_text(
+        main_path,
+        "ON ERROR DO handleerr WITH MESSAGE(), PROGRAM(), LINENO(), ERROR()\n"
+        "DO missing_routine\n"
+        "after_error = 1\n"
+        "RETURN\n"
+        "PROCEDURE handleerr\n"
+        "LPARAMETERS err_message, err_program, err_line, err_code\n"
+        "PUBLIC captured_message, captured_program, captured_line, captured_code\n"
+        "captured_message = err_message\n"
+        "captured_program = err_program\n"
+        "captured_line = err_line\n"
+        "captured_code = err_code\n"
+        "RETURN\n");
+
+    copperfin::runtime::PrgRuntimeSession session =
+        copperfin::runtime::PrgRuntimeSession::create(make_runtime_session_options(main_path.string(), temp_root.string(), false));
+
+    const auto state = session.run(copperfin::runtime::DebugResumeAction::continue_run);
+    expect(state.completed, "ON ERROR DO WITH script should complete after handler dispatch");
+
+    const auto captured_message = state.globals.find("captured_message");
+    const auto captured_program = state.globals.find("captured_program");
+    const auto captured_line = state.globals.find("captured_line");
+    const auto captured_code = state.globals.find("captured_code");
+    const auto after_error = state.globals.find("after_error");
+    expect(captured_message != state.globals.end(), "ON ERROR DO WITH handler should capture MESSAGE()");
+    expect(captured_program != state.globals.end(), "ON ERROR DO WITH handler should capture PROGRAM()");
+    expect(captured_line != state.globals.end(), "ON ERROR DO WITH handler should capture LINENO()");
+    expect(captured_code != state.globals.end(), "ON ERROR DO WITH handler should capture ERROR()");
+    expect(after_error != state.globals.end(), "execution should continue after ON ERROR DO WITH handler returns");
+
+    if (captured_message != state.globals.end()) {
+        expect(
+            copperfin::runtime::format_value(captured_message->second).find("Unable to resolve DO target") != std::string::npos,
+            "MESSAGE() should describe the failing command");
+    }
+    if (captured_program != state.globals.end()) {
+        expect(
+            copperfin::runtime::format_value(captured_program->second) == "main",
+            "PROGRAM() should report the failing procedure context");
+    }
+    if (captured_line != state.globals.end()) {
+        expect(
+            copperfin::runtime::format_value(captured_line->second) == "2",
+            "LINENO() should report the failing source line");
+    }
+    if (captured_code != state.globals.end()) {
+        expect(
+            copperfin::runtime::format_value(captured_code->second) == "1001",
+            "ERROR() should expose the first-pass runtime resolve error code");
+    }
+
+    const bool has_handler_event = std::any_of(
+        state.events.begin(),
+        state.events.end(),
+        [](const copperfin::runtime::RuntimeEvent& event) {
+            return event.category == "runtime.error_handler" &&
+                event.detail.find("WITH 4 argument") != std::string::npos;
+        });
+    expect(has_handler_event, "runtime should record ON ERROR DO WITH handler dispatch detail");
+
+    fs::remove_all(temp_root, ignored);
+}
+
+}  // namespace cf_test_prg_engine_control_flow
